requirements.txt
flask==3.0.0
authlib==1.3.0
requests==2.31.0
openpyxl==3.1.2
pandas==2.1.4
python-dotenv==1.0.0
gunicorn==21.2.0
